OneDrive/桌面/alpaca_trading_strategy/requirements.txt
alpaca-trade-api
alpaca-py
pandas
numpy
scipy
scikit-learn
matplotlib
seaborn
yfinance
pytz
vaderSentiment
aiohttp
tqdm
schedule
requests
python-dotenv